+++ test/Config/SentryGlobalConfigTest.php
@@ -36,6 +36,9 @@ final class SentryGlobalConfigTest extends TestCase
                 'max_depth'               => 8,
                 'max_items_per_container' => 100,
                 'max_total_nodes'         => 5000,
+                'use_default_rules'       => false,
+                'rules'                   => [],
+                'regex_rules'             => [],
             ],
             'http_context'                 => [
                 'enabled'               => true,